requirements.txt
Django~=3.2.9
django-filter~=21.1
django-ckeditor~=6.2.0
django-debug-toolbar~=3.2.3
